Initial Load of sketch and supporting files
Includes images of my implementation for reference in the docs/ folder.
Includes my blender file with the 3D models in the blender/ folder.
Includes my .stl model files I used in making the RiderCool Base in the
replicator/ folder.
Includes a non-XNP version of the sketch that can run the entire system.